requirements.txt
django==1.10
djangorestframework=3.4.6
django-widget-tweaks=1.4.1
numpy=1.11.1
pandas=0.18.1
scikit-learn==0.17.1
pickle